refactoring headers and fixing typos
